refactor: leaderboard repository
